Implement --syslog option
This option logs all commands to the system log before they are run;
for commands that fail, the kind of failure (exit with a status != 0,
killed by a signal, timeout) is logged as well.

The environment variable EXXE_IDENT can be used to change how exxe
identifies itself in the syslog.